Rename to _debug_mode.py to make it private
