+++ test/distributed/tensor/debug/test_debug_mode.py
@@ -14,8 +14,8 @@ from torch.testing._internal.common_utils import (
     TestCase,
 )
 from torch.testing._internal.distributed.fake_pg import FakeStore
+from torch.utils._debug_mode import DebugMode
 from torch.utils._python_dispatch import TorchDispatchMode
-from torch.utils.debug_mode import DebugMode
 
 
 @requires_cuda
